Updated webpack, babel, eslint. ES module build added. Split to separately importable pieces.
