+++ .babelrc
@@ -1,4 +1,37 @@
 {
-  "presets": ["es2015", "stage-0", "react"],
-  "plugins": ["transform-runtime", "transform-class-properties"]
+  "env": {
+    "commonjs": {
+      "presets": ["env"],
+      "plugins": [
+        "transform-runtime",
+        "transform-react-jsx",
+        "transform-class-properties",
+        "transform-object-rest-spread"
+      ]
+    },
+    "es": {
+      "presets": [
+        [
+          "env",
+          { "modules": false }
+        ]
+      ],
+      "plugins": [
+        "transform-runtime",
+        "transform-react-jsx",
+        "transform-class-properties",
+        "transform-object-rest-spread"
+      ]
+    },
+    "with_react_hot_loader": {
+      "presets": ["env"],
+      "plugins": [
+        "react-hot-loader/babel",
+        "transform-runtime",
+        "transform-react-jsx",
+        "transform-class-properties",
+        "transform-object-rest-spread"
+      ]
+    }
+  }
 }
