+++ .eslintignore
@@ -1,0 +1,1 @@
+demos/demo0/bundle.js
