+++ .eslintrc
@@ -1,21 +1,22 @@
 {
-  "extends": "eslint-config-airbnb",
+  "extends": "airbnb",
   "env": {
-    "es6": true
+    "es6": true,
+    "browser": true
   },
   "parser": "babel-eslint",
-  "ecmaFeatures": {
-    "modules": true,
-    "jsx": true
-  },
   "plugins": [
     "react"
   ],
   "rules": {
-    "react/prefer-es6-class": 0,
     "react/jsx-closing-bracket-location": 0,
-    "react/no-did-mount-set-state": 0,
-    "react/no-is-mounted": 0,
-    "react/prefer-stateless-function": 0
+    "react/jsx-first-prop-new-line": 0,
+    "react/require-default-props": 0,
+    "react/forbid-prop-types": 0,
+    "jsx-a11y/href-no-hash": 0,
+    "jsx-a11y/no-static-element-interactions": 0,
+    "no-underscore-dangle": 0,
+    "object-curly-newline": ["error", { "consistent": true }],
+    "prefer-template": 0
   }
 }
