+++ package.json
@@ -2,17 +2,19 @@
   "name": "react-share",
   "version": "1.19.0",
   "description": "Easy social media share buttons and share counts.",
-  "main": "./lib/react-share.js",
+  "main": "./lib/index.js",
+  "module": "./es/index.js",
   "engines": {
     "node": ">=4.0.0",
     "npm": ">=3.0.0"
   },
   "scripts": {
-    "build": "npm run lint && npm run build-src && npm run build-demos",
-    "build-src": "babel ./src --out-dir ./lib",
-    "build-demos": "NODE_ENV=production webpack -p --config webpack.demos.config.js",
-    "run-demos": "NODE_ENV=development webpack-dev-server --config webpack.demos.config.js --content-base demos/ --progress --colors",
-    "lint": "eslint --ext .jsx ./src ./demos",
+    "build": "npm run lint && npm run build:commonjs && npm run build:es && npm run build-demos",
+    "build:commonjs": "rimraf ./lib && cross-env BABEL_ENV=commonjs babel ./src --out-dir ./lib",
+    "build:es": "rimraf ./es && cross-env BABEL_ENV=es babel ./src --out-dir ./es",
+    "build-demos": "webpack -p --config webpack.demos.config.js",
+    "run-demos": "webpack-dev-server --hot --config webpack.demos.config.js --content-base demos/ --progress",
+    "lint": "eslint --ext .jsx --ext .js ./src ./demos",
     "prepublish": "npm run build",
     "start": "npm run run-demos"
   },
@@ -38,28 +40,30 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "babel-cli": "6.18.0",
-    "babel-core": "6.18.2",
-    "babel-eslint": "7.1.1",
-    "babel-loader": "6.2.10",
-    "babel-plugin-transform-class-properties": "6.18.0",
-    "babel-plugin-transform-runtime": "6.15.0",
-    "babel-preset-airbnb": "1.1.1",
-    "babel-preset-es2015": "6.18.0",
-    "babel-preset-react": "6.16.0",
-    "babel-preset-stage-0": "6.16.0",
-    "eslint": "2.13.1",
-    "eslint-config-airbnb": "7.0.0",
-    "eslint-loader": "1.6.3",
-    "eslint-plugin-jsx-a11y": "0.6.2",
-    "eslint-plugin-react": "4.3.0",
-    "file-loader": "0.9.0",
+    "babel-cli": "6.26.0",
+    "babel-core": "6.26.0",
+    "babel-eslint": "8.1.2",
+    "babel-loader": "7.1.2",
+    "babel-plugin-transform-class-properties": "6.24.1",
+    "babel-plugin-transform-object-rest-spread": "6.26.0",
+    "babel-plugin-transform-react-jsx": "^6.24.1",
+    "babel-plugin-transform-runtime": "6.23.0",
+    "babel-preset-env": "1.6.1",
+    "cross-env": "5.1.3",
+    "eslint": "4.14.0",
+    "eslint-config-airbnb": "16.1.0",
+    "eslint-loader": "1.9.0",
+    "eslint-plugin-import": "2.8.0",
+    "eslint-plugin-jsx-a11y": "6.0.3",
+    "eslint-plugin-react": "7.5.1",
+    "file-loader": "1.1.6",
     "react": "15.5.4",
     "react-dom": "15.5.4",
-    "react-hot-loader": "1.3.1",
-    "url-loader": "0.5.8",
-    "webpack": "1.13.3",
-    "webpack-dev-server": "1.16.3"
+    "react-hot-loader": "^3.1.3",
+    "rimraf": "^2.6.2",
+    "url-loader": "0.6.2",
+    "webpack": "3.10.0",
+    "webpack-dev-server": "2.9.7"
   },
   "dependencies": {
     "babel-runtime": "^6.6.1",
