--- src/react-share.jsx
@@ -1,13 +1,0 @@
-import { generateIcon } from './generateIcon';
-import * as ShareButtons from './share-buttons';
-import * as ShareCounts from './share-counts';
-import * as shareLinks from './social-media-share-links';
-import * as shareCountGetters from './share-count-getters';
-
-export {
-  generateIcon as generateShareIcon,
-  ShareButtons,
-  ShareCounts,
-  shareLinks,
-  shareCountGetters,
-};
