+++ src/utils/objectToGetParams.js
@@ -1,0 +1,8 @@
+/* eslint-disable prefer-template */
+export default function objectToGetParams(object) {
+  return '?' + Object.keys(object)
+    .filter(key => !!object[key])
+    .map(key => `${key}=${encodeURIComponent(object[key])}`)
+    .join('&');
+}
+/* eslint-enable prefer-template */
