+++ src/utils/shareCountFactory.jsx
@@ -1,18 +1,6 @@
-/* eslint-disable react/no-multi-comp */
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-
-import {
-  getFacebookShareCount,
-  getGooglePlusShareCount,
-  getLinkedinShareCount,
-  getPinterestShareCount,
-  getVKShareCount,
-  getOKShareCount,
-  getRedditShareCount,
-  getTumblrShareCount,
-} from './share-count-getters';
 
 class SocialMediaShareCount extends Component {
   constructor(props) {
@@ -43,7 +31,7 @@ class SocialMediaShareCount extends Component {
         isLoading: true,
       });
 
-      this.props.getCount(url, count => {
+      this.props.getCount(url, (count) => {
         if (this._isMounted) {
           this.setState({
             count,
@@ -84,21 +72,7 @@ SocialMediaShareCount.defaultProps = {
   children: shareCount => shareCount,
 };
 
-function shareCountFactory(getCount) {
+export default function shareCountFactory(getCount) {
   return props =>
     <SocialMediaShareCount getCount={getCount} {...props} />;
 }
-
-export const FacebookShareCount = shareCountFactory(getFacebookShareCount);
-export const LinkedinShareCount = shareCountFactory(getLinkedinShareCount);
-export const GooglePlusShareCount = shareCountFactory(getGooglePlusShareCount);
-export const PinterestShareCount = shareCountFactory(getPinterestShareCount);
-export const VKShareCount = shareCountFactory(getVKShareCount);
-export const OKShareCount = shareCountFactory(getOKShareCount);
-export const RedditShareCount = shareCountFactory(getRedditShareCount);
-export const TumblrShareCount = shareCountFactory(getTumblrShareCount);
-export const TwitterShareCount = () => {
-  throw new Error('TwitterShareCount was removed in version 1.3.0 because' +
-    'the Twitter API was shut down and there is no replacement. Please ' +
-    'remove it from your code.');
-};
